Adding support for implicit macro features.
Caveat:
If a macro is defined in the form

  <macro name="@MACRO">
    <lf>
      <satop nomvar="NOMVAR">
        <diamond mode="MODE">
          <prop name="NAME" />
        </diamond>
      </satop>
    </lf>
  </macro>

The resulting ccg entry looks like this:

MACRO<NOMVAR:MODE>: NAME;

However, ccg2xml drops the macro name and uses the prop name instead:

  <macro name="@NAME">
    <lf>
      <satop nomvar="NOMVAR">
        <diamond mode="MODE">
          <prop name="NAME" />
        </diamond>
      </satop>
    </lf>
  </macro>

This is functionaly equivalent, as the macro name is only an identifier.
However, this can lead to name clashes in certain circumstances as well
as issues with with unit tests (as the macro names now differ).